Keep files in a consistent order
Shows the selected image and keeps the files in order in the file list.

The filename is now saved from the uploaded file (and page numbers are added in the case of pdf). Saves a created timestamp for sorting.

Pulls file upload code out into one location.
